AOO41CG01...odt #137 Accepted Keith's changes + minor edits
Accepted all the edits except a few typos like a missing space or a
single letter deleted inside a word. I also added a comment about an
empty paragraph on page 37 that I believe is needed to preserve the
formatting of the heading that follows it. In some circumstatnces, a
heading immediately after a table or figure will get formatted to the
Text Body style when you save and reopen the document.